+++ src/MineClearance.Core/Interfaces/IMineGenerator.cs
@@ -10,8 +10,9 @@ internal interface IMineGenerator
     /// <summary>
     /// 生成地雷位置集合
     /// </summary>
-    /// <param name="config">对局配置</param>
+    /// <param name="config">游戏配置</param>
     /// <param name="firstClick">首次点击位置</param>
+    /// <param name="seed">随机种子, 用于生成固定的地雷布局</param>
     /// <returns>所有地雷位置的集合</returns>
-    IReadOnlyCollection<Models.Records.Position> GenerateMines(Models.Records.GameConfig config, Models.Records.Position firstClick);
+    IReadOnlyCollection<Models.Records.Position> GenerateMines(Models.Records.GameConfig config, Models.Records.Position firstClick, int seed);
 }